+++ docs/RELEASE-NOTES.md
@@ -1,1 +1,3 @@
-[2.3.1] Links between set meta
+[2.3.2] Testing to enforce chosen champions + fixes
+- VDecklist spacing + chosen champion right after legend
+- Added Red Kai'Sa's missing champion tag